CLN: tidyed unit tests and added probe_classes, factored out setup and review tests into seperate files
