+++ .gitignore
@@ -8,3 +8,4 @@ doc/build/
 *.json
 *.log
 *.csv
+.idea
