+++ force_wfmanager/tests/mock_methods.py
@@ -1,38 +1,10 @@
+import subprocess
 from unittest import mock
 
-from force_bdss.api import Workflow, WorkflowReader, WorkflowWriter
-from force_bdss.tests.probe_classes.factory_registry import (
-    ProbeFactoryRegistry
+from force_bdss.api import (
+    Workflow, WorkflowReader, WorkflowWriter,
+    InvalidFileException
 )
-
-from force_wfmanager.plugins.wfmanager_plugin import WfManagerPlugin
-from force_wfmanager.wfmanager_review_task import WfManagerReviewTask
-from force_wfmanager.wfmanager_setup_task import WfManagerSetupTask
-
-
-def mock_create_setup_task(filename):
-    def func():
-        wf_manager_task = WfManagerSetupTask(
-            factory_registry=ProbeFactoryRegistry())
-        if filename is not None:
-            wf_manager_task.load_workflow(filename)
-        return wf_manager_task
-    return func
-
-
-def mock_create_review_task():
-    def func():
-        wf_manager_task = WfManagerReviewTask(
-            factory_registry=ProbeFactoryRegistry())
-        return wf_manager_task
-    return func
-
-
-def mock_wfmanager_plugin(filename):
-    plugin = WfManagerPlugin()
-    plugin._create_setup_task = mock_create_setup_task(filename)
-    plugin._create_review_task = mock_create_review_task()
-    return plugin
 
 
 def mock_dialog(dialog_class, result, path=''):
@@ -59,3 +31,37 @@ def mock_file_writer(*args, **kwargs):
     writer = mock.Mock(spec=WorkflowWriter)
     writer.write = write
     return writer
+
+
+def mock_confirm_function(result):
+    def mock_conf(*args, **kwargs):
+        return result
+    return mock_conf
+
+
+def mock_file_reader_failure(*args, **kwargs):
+    def read(*args, **kwargs):
+        raise InvalidFileException("OUPS")
+    reader = mock.Mock(spec=WorkflowReader)
+    reader.read = read
+    return reader
+
+
+def mock_os_remove(*args, **kwargs):
+    raise OSError("OUPS")
+
+
+def mock_show_error(*args, **kwargs):
+    return args
+
+
+def mock_subprocess(*args, **kwargs):
+    def check_call(*args, **kwargs):
+        return
+    mock_subprocess_module = mock.Mock(spec=subprocess)
+    mock_subprocess_module.check_call = check_call
+    return mock_subprocess_module
+
+
+def mock_info_dialog(*args, **kwargs):
+    return args
