+++ force_wfmanager/ui/tests/test_ui_utils.py
@@ -2,9 +2,11 @@ import unittest
 
 from force_bdss.tests.dummy_classes.extension_plugin import \
     DummyExtensionPlugin
+from force_bdss.tests.probe_classes.data_source import \
+    ProbeDataSourceModel
 
 from force_wfmanager.ui.ui_utils import get_factory_name, model_info
-from force_wfmanager.tests.dummy_classes import DummyFactory, DummyModel
+from force_wfmanager.tests.dummy_classes import DummyFactory
 
 
 class TestUiUtils(unittest.TestCase):
@@ -13,7 +15,7 @@ class TestUiUtils(unittest.TestCase):
 
         self.plugin = DummyExtensionPlugin()
         self.factory = DummyFactory(self.plugin)
-        self.model = DummyModel(self.factory)
+        self.model = ProbeDataSourceModel(self.factory)
 
     def test_get_factory_name(self):
 
@@ -24,4 +26,4 @@ class TestUiUtils(unittest.TestCase):
     def test_model_info(self):
 
         self.assertEqual(model_info(None), [])
-        self.assertEqual(model_info(self.model), ['test_trait'])
+        self.assertEqual(len(model_info(self.model)), 4)
